Add fetch error logging and rename errors.js file
